fix plotting

+++ Python3-MODELS/Juan_Esteban/5-Stokes_Profiles_v1/try_sprof.py
@@ -14,7 +14,7 @@ def main():
         prof_im.append(mprof.read_prof(path+file, 'nicole',  nx, ny, nlam, ix, iy))
     fig, ax = plt.subplots(2, 2, figsize = (10,10))
     x = np.arange(0, len(prof_im[0]), 1)
-    for i in range(len(prof_im)/2):
+    for i in range(int(len(prof_im)/2)):
         ax[i%2, i].scatter(x,prof_im[i])
     fig.savefig("first_prof.png")
